+++ rmap/src/common/load_reads.hpp
@@ -34,32 +34,32 @@
 void
 load_reads_from_fasta_file(const std::string &filename, const size_t max_diffs,
 			   size_t &read_width, std::vector<FastRead> &fast_reads,
-			   std::vector<size_t> &read_words, std::vector<size_t> &read_index);
+			   std::vector<size_t> &read_words, std::vector<unsigned int> &read_index);
 
 void
 load_reads_from_fastq_file(const std::string &filename, const size_t max_diffs,
 			   size_t &read_width, std::vector<FastRead> &fast_reads,
-			   std::vector<size_t> &read_words, std::vector<size_t> &read_index);
+			   std::vector<size_t> &read_words, std::vector<unsigned int> &read_index);
 
 void
 load_reads_from_fastq_file(const std::string &filename, const size_t max_diffs,
 			   size_t &read_width, std::vector<FastReadWC> &fast_reads,
-			   std::vector<size_t> &read_words, std::vector<size_t> &read_index);
+			   std::vector<size_t> &read_words, std::vector<unsigned int> &read_index);
 
 void
 load_reads_from_fastq_file(const std::string &filename, const size_t max_diffs,
 			   size_t &read_width, std::vector<FastReadQuality> &fast_reads,
-			   std::vector<size_t> &read_words, std::vector<size_t> &read_index);
+			   std::vector<size_t> &read_words, std::vector<unsigned int> &read_index);
 
 void
 load_reads_from_prb_file(const std::string &filename, const size_t max_diffs,
 			 size_t &read_width, std::vector<FastReadWC> &fast_reads,
-			 std::vector<size_t> &read_words, std::vector<size_t> &read_index);
+			 std::vector<size_t> &read_words, std::vector<unsigned int> &read_index);
 
 void
 load_reads_from_prb_file(const std::string &filename, const size_t max_diffs,
 			 size_t &read_width, std::vector<FastReadQuality> &fast_reads,
-			 std::vector<size_t> &read_words, std::vector<size_t> &read_index);
+			 std::vector<size_t> &read_words, std::vector<unsigned int> &read_index);
 
 
 #endif
